% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{books.insert.callback}
\alias{books.insert.callback}
\title{新增}
\usage{
books.insert.callback(data, row, table = "t_tsp_ques", f = getBooks,
  id_var = "FId")
}
\arguments{
\item{data}{数据}

\item{row}{行}

\item{table}{列表}

\item{f}{函数}

\item{id_var}{内码}
}
\value{
返回值
}
\description{
新增
}
\examples{
books.insert.callback()
}
