% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{books.update.callback}
\alias{books.update.callback}
\title{更新数据}
\usage{
books.update.callback(
  data,
  olddata,
  row,
  table = "t_tsp_ques",
  f = getBooks,
  edit.cols = c("FQues", "FAnsw"),
  id_var = "FId"
)
}
\arguments{
\item{data}{新数据}

\item{olddata}{原数据}

\item{row}{行}

\item{table}{表}

\item{f}{函数}

\item{edit.cols}{列名}

\item{id_var}{内码}
}
\value{
返回值
}
\description{
更新数据
}
\examples{
books.update.callback()
}
