% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{books.update.callback2}
\alias{books.update.callback2}
\title{更校报2}
\usage{
books.update.callback2(
  data,
  olddata,
  row,
  table = "t_tsp_ques",
  f = getBooks2,
  edit.cols = c("FQues", "FAnsw"),
  id_var = "FId"
)
}
\arguments{
\item{data}{新数据}

\item{olddata}{原数据}

\item{row}{行}

\item{table}{表}

\item{f}{函数}

\item{edit.cols}{列}

\item{id_var}{内码}
}
\value{
返回值
}
\description{
更校报2
}
\examples{
books.update.callback2()
}
