% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{csp_books.update.callback}
\alias{csp_books.update.callback}
\title{处理更新事项}
\usage{
csp_books.update.callback(
  data,
  olddata,
  row,
  table = "t_tsp_ques",
  f = csp_getBooks,
  edit.cols = c("FQues", "FAnsw"),
  id_var = "FId",
  FCspName = "巴豆"
)
}
\arguments{
\item{data}{数据}

\item{olddata}{原来数据}

\item{row}{新的数据}

\item{table}{表}

\item{f}{函数}

\item{edit.cols}{列}

\item{id_var}{内码}

\item{FCspName}{客服}
}
\value{
返回值
}
\description{
处理更新事项
}
\examples{
csp_books.update.callback()
}
