% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{csp_books.update.callback2}
\alias{csp_books.update.callback2}
\title{数据}
\usage{
csp_books.update.callback2(
  data,
  olddata,
  row,
  table = "t_tsp_ques",
  f = csp_getBooks2,
  edit.cols = c("FQues", "FAnsw"),
  id_var = "FId",
  FCspName = "巴豆"
)
}
\arguments{
\item{data}{处理}

\item{olddata}{数据}

\item{row}{行}

\item{table}{表}

\item{f}{函数}

\item{edit.cols}{个签}

\item{id_var}{经}

\item{FCspName}{在}
}
\value{
返回值
}
\description{
数据
}
\examples{
csp_books.update.callback2()
}
