% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{csp_getBooks}
\alias{csp_getBooks}
\title{处理内容}
\usage{
csp_getBooks(table = "t_tsp_ques", FCspName = "巴豆")
}
\arguments{
\item{table}{表}

\item{FCspName}{客服名称}
}
\value{
返回值
}
\description{
处理内容
}
\examples{
csp_getBooks()
}
