% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{csp_getBooks2}
\alias{csp_getBooks2}
\title{插入值}
\usage{
csp_getBooks2(table = "t_tsp_ques", FCspName = "巴豆")
}
\arguments{
\item{table}{表}

\item{FCspName}{客服名称}
}
\value{
返回值
}
\description{
插入值
}
\examples{
csp_getBooks()
}
