% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getMax_id}
\alias{getMax_id}
\title{获取最大ID}
\usage{
getMax_id(conn, table = "t_tsp_ques", id_var = "FId")
}
\arguments{
\item{conn}{连接}

\item{table}{表名}

\item{id_var}{内码}
}
\value{
返回值
}
\description{
获取最大ID
}
\examples{
getMax_id()
}
