% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icLog.R
\name{icLogUpload}
\alias{icLogUpload}
\title{写入客服操作日志}
\usage{
icLogUpload(
  conn = conn_rds_nsic(),
  FNickName = "RDS",
  FQuesText = "发现运动版多少钱",
  answ,
  index = 1,
  type = "A"
)
}
\arguments{
\item{conn}{数据库连接}

\item{FNickName}{客服名称，默认为登录名}

\item{FQuesText}{原始查询问题}

\item{answ}{问题回复数据包}

\item{index}{可中部分}

\item{type}{问题序号}
}
\value{
返回值
}
\description{
写入客服操作日志
}
\examples{
icLogUpload()
}
