% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icLog.R
\name{queryLog_upload}
\alias{queryLog_upload}
\title{写入AI操作日志功能}
\usage{
queryLog_upload(
  conn = conn_rds_nsic(),
  FNickName = "RDS",
  FQuesText = "发现运动版多少钱",
  answ
)
}
\arguments{
\item{conn}{数据库连接}

\item{FNickName}{客服名称}

\item{FQuesText}{问题名}

\item{answ}{答案}
}
\value{
返回值
}
\description{
写入AI操作日志功能
}
\examples{
queryLog_upload()
}
