% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ques_commit}
\alias{ques_commit}
\title{问题提交}
\usage{
ques_commit(
  FCspName = "bot19",
  FQues = "多少钱",
  FAnsw = "请内部支持",
  FTspName = "badou"
)
}
\arguments{
\item{FCspName}{客服名称}

\item{FQues}{问题}

\item{FAnsw}{答案}

\item{FTspName}{支持人员}
}
\value{
返回值
}
\description{
问题提交
}
\examples{
ques_commit()
}
